pyproject.toml
[tool.poetry]
name = "open-ai-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.13.3"
streamlit = "^1.32.0"
pygithub = "^2.2.0"
python-dotenv = "^1.0.1"
black = "^24.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
